Actual modification of the spreadsheet through the application
You can now modify the spreadsheet through the application, you can access the page to modifi a profile by consultic it in the search page.
For now it's for every peoples registered, but it's a probleme for another time.
(Sorry for the grammatical errors, should i really continue to write in english)